fixed foot note action
