+++ src/redux/actions/media.js
@@ -2,7 +2,9 @@ import {
   SET_MEDIA,
   REMOVE_MEDIA,
   LOAD_TAFISRS,
-  LOAD_TAFISRS_SUCCESS
+  LOAD_TAFISRS_SUCCESS,
+  LOAD_FOOT_NOTE,
+  LOAD_FOOT_NOTE_SUCCESS
 } from 'redux/constants/media';
 
 export const setMedia = content => ({
@@ -18,4 +20,9 @@ export const loadTafsirs = verse => ({
   types: [LOAD_TAFISRS, LOAD_TAFISRS_SUCCESS],
   promise: client => client.get('/api/v3/options/tafsirs'),
   verse
+});
+
+export const loadFootNote = footNoteId => ({
+  types: [LOAD_FOOT_NOTE, LOAD_FOOT_NOTE_SUCCESS],
+  promise: client => client.get(`/api/v3/foot_notes/${footNoteId}`)
 });
