--- src/redux/constants/footNote.js
@@ -1,3 +1,0 @@
-export const LOAD_FOOT_NOTE = '@@quran/footNote/LOAD';
-export const LOAD_FOOT_NOTE_SUCCESS = '@@quran/footNote/LOAD_SUCCESS';
-export const LOAD_FOOT_NOTE_FAIL = '@@quran/footNote/LOAD_FAIL';
