+++ src/redux/constants/media.js
@@ -3,3 +3,5 @@ export const SET_OPTIONS = '@@quran/auth/SET_OPTIONS';
 export const REMOVE_MEDIA = '@@quran/auth/REMOVE_MEDIA';
 export const LOAD_TAFISRS = '@@quran/popup/LOAD_TAFISRS';
 export const LOAD_TAFISRS_SUCCESS = '@@quran/popup/LOAD_TAFISRS_SUCCESS';
+export const LOAD_FOOT_NOTE = '@@quran/popup/LOAD_FOOT_NOTE';
+export const LOAD_FOOT_NOTE_SUCCESS = '@@quran/popup/LOAD_FOOT_NOTE_SUCCESS';
