+++ src/styles/main.scss
@@ -171,7 +171,7 @@ li em{
   line-height: 30px;
 
   #arabic{
-    font-family: initial;
+    font-family: monospace;
   }
 }
